Finish basic implementation of PTP/IP - tested on Fujifilm X-A2 - Rename some functions to be less wordy - wifi.c test file connects, gets device ready for reading, then pings until shutdown
Tried to get an event pipe working, but couldn't figure it out. It's probably on another port,
but the app doesn't even seem to connect to it. Could not even be implemented on the camera itself.
Constant pinging is required to keep the connection alive (and check if it pings back), so a dummy
pinger will do for now. (ptp_fuji_ping)

PTP/IP events are still sent through the command socket, so I modified the backend reciever to check
for the shutdown event. This should work for now, and probably won't cause any issues.